fix: invite leaderboard

+++ src/modules/moderate.js
@@ -9,13 +9,14 @@ const moderate = (bot) => {
     await wait(300);
     const guild = await bot?.guilds?.fetch(process.env.DISCORD_GUILD_ID);
     const member = await guild?.members?.fetch(message.author.id);
-    //const isAdmin = member?.roles?.cache?.has(
-    // process.env.DISCORD_ADMIN_ROLE_ID
-    // );
+    const isAdmin = member?.roles?.cache?.has(
+      process.env.DISCORD_ADMIN_ROLE_ID
+    );
 
     if (
       message.channelId === process.env.DISCORD_COMMAND_ONLY_CHANNEL &&
-      message.interaction?.type !== "APPLICATION_COMMAND"
+      message.interaction?.type !== "APPLICATION_COMMAND" &&
+      !isAdmin
     ) {
       message.delete();
     }
